+++ src/app/app-routing.module.ts
@@ -3,7 +3,7 @@ import { Routes, RouterModule } from '@angular/router';
 import { MainPageComponent } from './main-page/main-page.component';
 import { WawbComponent } from './maps/m1wawb/wawb.component';
 import { TroubledwatersComponent } from './maps/m3troubledwaters/troubledwaters.component';
-import { M6Component } from './maps/m6/m6.component';
+import { M6Component } from './maps/m6/m6-component';
 
 const routes: Routes = [
   {path: 'water-above-water-below', component: WawbComponent},
