+++ src/app/maps/m6/directives/m6-arrowpre.directive.spec.ts
@@ -1,4 +1,4 @@
-import { ArrowpreDirective } from './arrowpre.directive';
+import { ArrowpreDirective } from './m6-arrowpre.directive';
 
 describe('ArrowpreDirective', () => {
   it('should create an instance', () => {
