+++ src/app/maps/m6/directives/m6-next.directive.spec.ts
@@ -1,4 +1,4 @@
-import { NextDirective } from './next.directive';
+import { NextDirective } from './m6-next.directive';
 
 describe('NextDirective', () => {
   it('should create an instance', () => {
