+++ src/app/maps/m6/directives/m6-pagination.directive.spec.ts
@@ -1,4 +1,4 @@
-import { PaginationDirective } from './pagination.directive';
+import { PaginationDirective } from './m6-pagination.directive';
 
 describe('PaginationDirective', () => {
   it('should create an instance', () => {
