+++ src/app/maps/m6/directives/m6-prev.directive.spec.ts
@@ -1,4 +1,4 @@
-import { PrevDirective } from './prev.directive';
+import { PrevDirective } from './m6-prev.directive';
 
 describe('PrevDirective', () => {
   it('should create an instance', () => {
