+++ src/app/maps/m6/directives/m6-shortcut.directive.spec.ts
@@ -1,4 +1,4 @@
-import { ShortcutDirective } from './shortcut.directive';
+import { ShortcutDirective } from './m6-shortcut.directive';
 
 describe('ShortcutDirective', () => {
   it('should create an instance', () => {
