+++ src/app/maps/m6/directives/m6-shortcut.directive.ts
