+++ src/app/maps/m6/directives/m6-synch.directive.ts
