+++ src/app/maps/m6/info/m6-info.component.less
