+++ src/app/maps/m6/info/m6-info.component.ts
@@ -3,11 +3,11 @@ import { EventEmitter } from '@angular/core';
 import { Component, OnInit } from '@angular/core';
 
 import * as marked from 'marked';
-import { M6Service } from '../m6.service';
+import { M6Service } from '../m6-service';
 @Component({
   selector: 'app-info',
-  templateUrl: './info.component.html',
-  styleUrls: ['./info.component.less']
+  templateUrl: './m6-info.component.html',
+  styleUrls: ['./m6-info.component.less']
 })
 export class InfoComponent implements OnInit {
