+++ src/app/maps/m6/m6-component.less
