+++ src/app/maps/m6/model/pagination/m6-pagination.component.less
