+++ src/app/maps/m6/model/pagination/m6-pagination.component.ts
@@ -2,8 +2,8 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'pagination',
-  templateUrl: './pagination.component.html',
-  styleUrls: ['./pagination.component.less']
+  templateUrl: './m6-pagination.component.html',
+  styleUrls: ['./m6-pagination.component.less']
 })
   
 export class PaginationComponent {}
