+++ src/app/maps/m6/seclayout/m6-seclayout.component.less
