+++ src/app/maps/m6/seclayout/m6-seclayout.component.ts
@@ -4,8 +4,8 @@ import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 
 @Component({
   selector: 'app-seclayout',
-  templateUrl: './seclayout.component.html',
-  styleUrls: ['./seclayout.component.less']
+  templateUrl: './m6-seclayout.component.html',
+  styleUrls: ['./m6-seclayout.component.less']
 })
 export class SeclayoutComponent implements OnInit {
   @Input() hideHeader = false;
